be/requirements.txt
# Flask Core
flask>=3.0.0
python-dotenv>=1.0.0

# Database
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.9

# Migrations
flask-migrate>=4.0.0

# Authentication
flask-jwt-extended>=4.6.0
bcrypt>=4.1.0

# Validation
marshmallow>=3.20.0

# API Utils
flask-cors>=4.0.0
flask-limiter>=3.5.0

# Email
flask-mail>=0.9.1

# File Upload
cloudinary>=1.36.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Development
gunicorn>=21.2.0
